[ALPS07433092] scsi: ufs: fix coverity issues
CIDs
- 11361579 ufs_uic_pa_err_str[bit] evaluates to an address that could
    be at negative offset of an array
- 11361644 Invalid type in argument to printf format specifier
- 11361551 Dereference before null check

MTK-Commit-Id: 2ebcf8c3a2d0577d496cf7170d7b116acab21362

CR-Id: ALPS07433092
Feature: ufs

+++ drivers/scsi/ufs/ufs-mediatek.c
@@ -2407,7 +2407,7 @@ static void ufs_mtk_event_notify(struct ufs_hba *hba,
 {
 	unsigned int val = *(u32 *)data;
 	unsigned long reg;
-	int bit;
+	uint8_t bit;
 
 	trace_ufs_mtk_event(evt, val);
